feat: four probe presets, verified against the seeded defects
Presets are builders, not fixed probes: they read endpoints and request
params from config so they adapt to the service under test.

Integration tests run the real target app in-process and assert each
preset catches its defect and stays quiet on healthy endpoints.

Three defects found in my own probe code while writing those tests:
- percentile() contradicted its "nearest-rank" docstring, returning an
  interpolated index. Now ceil(q*n), which always returns an observed value.
- the template-injection check flagged any body containing "49" (from 7*7).
  Replaced with a differential check against a control response using an
  operand product that cannot occur naturally.
- load and SLO presets sent no query params, so they never reached the
  buggy code path they were meant to exercise.

+++ services/control-plane/sandman/verdict.py
@@ -118,7 +118,7 @@ class LaneSummary:
         if not self.latencies_ms:
             return None
         ordered = sorted(self.latencies_ms)
-        idx = min(len(ordered) - 1, int(round(0.95 * (len(ordered) - 1))))
+        idx = min(len(ordered) - 1, round(0.95 * (len(ordered) - 1)))
         return ordered[idx]
 
     def add(self, result: ProbeResult) -> None:
